+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 - Minification safe fixes
 - Adds debug testing tool pkg.scripts.test:debug
 - (Internal) enums no longer used due to awkwardness between TS enums and Google Closure Compiler enums
+- No more global state for responders
+- No more global uid state
+- Faster uid function
+- exports as es6
+- does not export as goog.module
 
 ## 1.0.2